+++ sqlalchemy_utils/types/__init__.py
@@ -9,7 +9,8 @@ from .country import CountryType  # noqa
 from .currency import CurrencyType  # noqa
 from .email import EmailType  # noqa
 from .encrypted.encrypted_type import EncryptedType  # noqa
-from .enriched_datetime import EnrichedDateTimeType, EnrichedDateType  # noqa
+from .enriched_datetime.enriched_datetime import EnrichedDateTimeType  # noqa
+from .enriched_datetime.enriched_datetime import EnrichedDateType  # noqa
 from .ip_address import IPAddressType  # noqa
 from .json import JSONType  # noqa
 from .locale import LocaleType  # noqa
